Attempting to resolve travis CI error
I had to make changes to the spark and spark-dependencies pom's to get spark.version related defined variables to propogate properly.  This created an issue when attempting to build against Spark-1.4.  For integration with the master, the poms are going to have to be integrated.  I don't feel I'm in a position to propose how to do that, because there are simply too many moving parts to the zeppelin-spark build sequence and it changes too often.  For testing purposes, I've therefore taken 1.4 and prior out of the travis script.